Implement haversine distance formula
Calculates distance using either a pair of matrix of lat-long
values or vector inputs for each pair of lat and long values.

Input values are in radians.

Haversine distance calculates the great circle distance between
two points on a sphere of given radius, which is 1 by default.

It is useful here to calculate the approximate mileage between two
points on earth's surface.