Remove uneccesary function call in randwalk
It is not necessary to call any because if any of the elements are
not false, it will result in a vector that isn't all true, so the
while loop will stop.  In other words, any is necessary to change
the stop condition of a while loop to be `all`, but since we are
negating the logical vector result of the stop function, the call
to any is redundant.