Improve responsive logic for visualizations
Better adapt to screen resolution, and will now, given a small
screen size, grow to fit the width of the screen, and force the
flex-box display to wrap to columns.

Reduce the number of cells that appear on the screen from 20 in
the dominant direction, to 16 in the dominant direction to improve
performance of the algorithm.