Refactor sketch initialization
Create main.js file which seperates each sketch into an object
specifying the sketch function and parent element id, and then
maps over the set of sketches, passing them as arguments to the
new p5() function.

Cleaner and easier to edit, easier to change initialization logic
that will effect every sketch, and provides a sensible place to
declare constants that will be used by the sketches.