Fix priority queue usage in AStar and Dijkstra
To get the proper behaviour, we need to increase the priority of
nodes if we find a better path to them.  Since the js priority
queue implementation has no such procedure, simply re-insert the
search item and skip it when popping the queue if it has already
been visited before, ensuring that if we find an element a second
time from the queue, we don't waste an iteration on it.